+++ app/Http/Controllers/HomeController.php
@@ -3,13 +3,33 @@
 namespace App\Http\Controllers;
 
 use App\Article;
+use App\Authority;
 use App\Collection;
+use App\FeaturedArtwork;
+use App\FeaturedPiece;
 
 class HomeController extends Controller
 {
     public function index()
     {
         // TODO caching
+        $featuredPiece = FeaturedPiece::query()
+            ->published()
+            ->with('media')
+            ->orderBy('updated_at', 'desc')
+            ->first();
+
+        $featuredArtwork = FeaturedArtwork::query()
+            ->published()
+            ->orderBy('updated_at', 'desc')
+            ->first();
+
+        $featuredAuthor = Authority::query()
+            ->where('has_image', true)
+            ->where('type', 'person')
+            ->has('items', '>', 3)
+            ->inRandomOrder()
+            ->first();
 
         $articles = Article::query()
             ->with(['translations', 'category'])
@@ -32,11 +52,16 @@ class HomeController extends Controller
         $collectionsTotalCount = Collection::published()->count();
         $collectionsRemainingCount = floor(($collectionsTotalCount - 5) / 10) * 10; // Round down to nearest 10
 
-        return view('home.index')->with(compact([
-            'articles',
-            'articlesRemainingCount',
-            'collections',
-            'collectionsRemainingCount'
-        ]));
+        return view('home.index')->with(
+            compact([
+                'featuredPiece',
+                'featuredArtwork',
+                'featuredAuthor',
+                'articles',
+                'articlesRemainingCount',
+                'collections',
+                'collectionsRemainingCount',
+            ])
+        );
     }
 }