+++ app/Item.php
@@ -5,11 +5,15 @@
 namespace App;
 
 use App\Contracts\IndexableModel;
+use App\Elasticsearch\Repositories\ItemRepository;
 use App\Events\ItemPrimaryImageChanged;
 use App\Matchers\AuthorityMatcher;
-use Chelout\RelationshipEvents\Concerns\HasBelongsToManyEvents;
 use Astrotomic\Translatable\Translatable;
 use Astrotomic\Translatable\Contracts\Translatable as TranslatableContract;
+use Chelout\RelationshipEvents\Concerns\HasBelongsToManyEvents;
+use ElasticScoutDriverPlus\Builders\BoolQueryBuilder;
+use ElasticScoutDriverPlus\Searchable;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Date;
@@ -24,8 +28,10 @@ use Symfony\Component\Validator\Mapping\ClassMetadata;
 class Item extends Model implements IndexableModel, TranslatableContract
 {
     use \Conner\Tagging\Taggable;
-    use Translatable;
     use HasBelongsToManyEvents;
+    use HasFactory;
+    use Searchable;
+    use Translatable;
 
     const ARTWORKS_DIR = '/images/diela/';
 
@@ -38,6 +44,29 @@ class Item extends Model implements IndexableModel, TranslatableContract
     const COLOR_AMOUNT_THRESHOLD = 0.03;
 
     protected $keyType = 'string';
+
+    public static $filterables = [
+        'author',
+        'topic',
+        'work_type',
+        'medium',
+        'technique',
+        'gallery',
+        'additionals.category.keyword',
+        'additionals.frontend.keyword',
+        'additionals.set.keyword',
+        'additionals.location.keyword',
+    ];
+
+    public static $rangeables = [
+        'date_earliest',
+        'date_latest',
+        'additionals.order',
+    ];
+
+    public static $sortables = [
+        'additionals.order',
+    ];
 
     public $translatedAttributes = [
         'title',
@@ -129,6 +158,8 @@ class Item extends Model implements IndexableModel, TranslatableContract
         'belongsToManyUpdatingExistingPivot',
         'belongsToManyUpdatedExistingPivot',
     ];
+
+    protected $appends = ['image_url'];
 
     protected $useTranslationFallback;
 
@@ -742,5 +773,34 @@ class Item extends Model implements IndexableModel, TranslatableContract
     public function setUseTranslationFallback(?bool $useTranslationFallback)
     {
         $this->useTranslationFallback = $useTranslationFallback;
+    }
+
+    public function searchableAs()
+    {
+        return app(ItemRepository::class)->getLocalizedIndexName();
+    }
+
+    public function getImageUrlAttribute()
+    {
+        return sprintf('%s%s', config('app.url'), $this->getImagePath());
+    }
+
+    public static function filterQuery(array $filter, BoolQueryBuilder $builder = null)
+    {
+        $builder = $builder ?: new BoolQueryBuilder();
+        foreach ($filter as $field => $value) {
+            if (is_string($value) && in_array($field, self::$filterables, true)) {
+                $builder->filter('term', [$field => $value]);
+            } else if (is_array($value) && in_array($field, self::$rangeables, true)) {
+                $range = collect($value)
+                    ->only(['lt', 'lte', 'gt', 'gte'])
+                    ->transform(function ($value) {
+                        return (string)$value;
+                    })
+                    ->all();
+                $builder->filter('range', [$field => $range]);
+            }
+        }
+        return $builder;
     }
 }