+++ resources/js/zoom.js
@@ -1,0 +1,2 @@
+window.$ = window.jQuery = require('jquery');
+window.OpenSeadragon = require('openseadragon');