+++ resources/views/components/admin/link.blade.php
@@ -1,1 +1,1 @@
-<a {!! $attributes->merge(['class' => 'tw-text-[#428bca] hover:tw-underline']) !!}>{{ $slot }}</a>
+<a {!! $attributes->merge(['class' => 'tw-text-[#428bca] hover:tw-text-[#428bca] hover:tw-underline']) !!}>{{ $slot }}</a>