+++ resources/views/zoom.blade.php
@@ -20,9 +20,6 @@
 
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
   <link rel="stylesheet" type="text/css" href="{{ mix('/css/style.css') }}" />
-
-  {{-- JS --}}
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.10.2/jquery.min.js"></script>
 </head>
 <body class="template-zoom">
   @include('components.zoomviewer', [
@@ -30,6 +27,10 @@
     'index' => $index,
     'fullIIPImgURLs' => $fullIIPImgURLs,
   ])
+
+  <script type="text/javascript" src="{{ mix('/js/manifest.js') }}"></script>
+  <script type="text/javascript" src="{{ mix('/js/vendor.js') }}"></script>
+  <script type="text/javascript" src="{{ mix('/js/zoom.js') }}"></script>
 
   @yield('javascript')
 </body>